Initial commit. Added compact perl executable.
Pretty much feature complete, a lot of great features. Perl is awesome!
I still want to add a --delimiter option so you can delimit your entries
with something other than a newline if you want to. Also the naming of
some of the variables could possibly be improved. It works great! The
only reason it's not ready for production is that there is a TODO in the
documentation left that I still have to do. There can of course still be
bugs in it, but time will tell.